Hack a browser web worker build.
I've seen it give syntax errors and basic completions.

No types are available except the one that I've rammed into the file
system as a proof of concept.

There are a couple of changes outside of the new package that need some
careful consideration.

Lots could be broken here as a variety of NodeJS modules have been
stubbed out or polyfilled.