Change default to use project root ".venv"
fixes #365
If project root is found, and ".venv" is found under project root,
use that venv as default.

Project root is found if "pyrightconfig.json" or "pyproject.toml" is
found. ".venv" can change to another name by change the
'[tool.pyright] venv="venv_name"' in "pyproject.toml".
But note that `uv` only support ".venv" so better stick with that.